drawsettings enhanced
git-svn-id: svn+ssh://erebor.esa.informatik.tu-darmstadt.de/home/wimi/svn/GaalopCompiler/gaalop-2.0/branches/tbaAndGapp@1153 4235ee25-d3ef-403b-b6ed-692c26853d4b

+++ codegen-visualizer/src/main/java/de/gaalop/visualizer/zerofinding/RayMethodThread.java
@@ -17,7 +17,7 @@ public class RayMethodThread extends Thread {
     private float fromOY_Incl;
     private float toOY_Excl;
     
-    private int a ;
+    private float a ;
     private float dist;
     private ControlFlowGraph graph;
     
@@ -28,7 +28,7 @@ public class RayMethodThread extends Thread {
     
     public HashMap<String, LinkedList<Point3d>> points = new HashMap<String, LinkedList<Point3d>>();
 
-    public RayMethodThread(float fromOY_Incl, float toOY_Excl, int a, float dist, ControlFlowGraph graph, HashMap<MultivectorComponent, Double> globalValues, LinkedList<AssignmentNode> nodes) {
+    public RayMethodThread(float fromOY_Incl, float toOY_Excl, float a, float dist, ControlFlowGraph graph, HashMap<MultivectorComponent, Double> globalValues, LinkedList<AssignmentNode> nodes) {
         this.fromOY_Incl = fromOY_Incl;
         this.toOY_Excl = toOY_Excl;
         this.a = a;
